third_party/WebKit/LayoutTests/http/tests/security/powerfulFeatureRestrictions/webcrypto-on-secure-origin-in-insecure-origin.html

<!DOCTYPE html>
<html>
<head>
<title>WebCrypto On A Secure Origin Embedded in an Insecure Origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
</head>
<body></body>
<script>
    if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
        window.location = get_host_info().UNAUTHENTICATED_ORIGIN +
                          window.location.pathname;
    } else {
        if (window.testRunner)
            testRunner.overridePreference(
                "WebKitStrictPowerfulFeatureRestrictions", true);

        async_test(function () {
            window.addEventListener("message", this.step_func(function (event) {
                assert_equals(event.data.success, true);
                this.done();
            }));
            var iframe = document.createElement("iframe");
            iframe.src = get_host_info().HTTP_ORIGIN +
                         "/security/powerfulFeatureRestrictions/resources/" +
                         "webcrypto.html";
            document.body.appendChild(iframe);
        }, "WebCrypto allowed on secure origin in insecure origin");
    }
</script>
</html>
